requirements.txt
asgiref==3.5.2
Django==4.1.2
django-filter==22.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
Markdown==3.4.1
pytz==2022.4
sqlparse==0.4.3
tzdata==2022.4
